Separate DataElements for clarity
End of day checkin
 
TODO fix unit tests
