+++ test/Adapter/MTCAdapterTests.cs
@@ -23,6 +23,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading;
+using MTConnect.DataElements;
 
 namespace MTConnect.utests.Adapter
 {
