+++ test/ConditionTests.cs
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
+using MTConnect.DataElements;
 using NUnit.Framework;
 using System.Collections.Generic;
