+++ test/DataItemTests.cs
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-using MTConnect;
+using MTConnect.DataElements;
 using NUnit.Framework;
 
 namespace MTConnect.utests
